+++ src/components/layout/customer/components/incomes/ApplyRuleModalContent.tsx
@@ -147,7 +147,7 @@ function ApplyRuleModalContent({
   return (
     <div className="space-y-8">
       <h1 className="font-medium text-xl text-[#000] mt-6 mb-8">
-        {translate('applyRuleModal.title')}
+        {translate('applyRuleModal.title_for_income')}
       </h1>
       <div className="flex flex-wrap gap-2">
         {incomesWithRules.map((incomeRule) => (
